+++ src/shared/components/results/styles/results-view.scss
@@ -1,4 +1,4 @@
-@import '../../../../shared/styles/settings';
+@import '../../../styles/settings';
 
 .is-stale {
   opacity: 0.5;
