+++ src/shared/components/results/styles/warning.scss
@@ -1,0 +1,3 @@
+.warning {
+  color: red;
+}
