--- src/uniprotkb/components/results/styles/warning.scss
@@ -1,3 +1,0 @@
-.warning {
-    color: red;
-}
